+++ lib/src/json/zones/zone.g.dart
@@ -16,6 +16,9 @@ Zone _$ZoneFromJson(Map<String, dynamic> json) => Zone(
       music: json['music'] == null
           ? null
           : Sound.fromJson(json['music'] as Map<String, dynamic>),
+      ambiances: (json['ambiances'] as List<dynamic>?)
+          ?.map((e) => Sound.fromJson(e as Map<String, dynamic>))
+          .toList(),
       topDownMap: json['topDownMap'] as bool? ?? true,
       edgeCommand: json['edgeCommand'] == null
           ? null
@@ -35,6 +38,7 @@ Map<String, dynamic> _$ZoneToJson(Zone instance) => <String, dynamic>{
       'boxes': instance.boxes,
       'defaultTerrainId': instance.defaultTerrainId,
       'music': instance.music,
+      'ambiances': instance.ambiances,
       'topDownMap': instance.topDownMap,
       'edgeCommand': instance.edgeCommand,
       'turnAmount': instance.turnAmount,